// src/Polyglot.Frontend/src/app/api/model/creditAdjustmentMode.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const CreditAdjustmentMode = {
    Add: 'Add',
    Remove: 'Remove',
    Overwrite: 'Overwrite'
} as const;
export type CreditAdjustmentMode = typeof CreditAdjustmentMode[keyof typeof CreditAdjustmentMode];