// src/Polyglot.Frontend/src/app/api/model/modelListMode.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const ModelListMode = {
    None: 'None',
    Whitelist: 'Whitelist',
    Blacklist: 'Blacklist'
} as const;
export type ModelListMode = typeof ModelListMode[keyof typeof ModelListMode];