// src/Polyglot.Frontend/src/app/api/model/modelListType.ts
/**
 * Polyglot.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const ModelListType = {
    Whitelist: 'Whitelist',
    Blacklist: 'Blacklist'
} as const;
export type ModelListType = typeof ModelListType[keyof typeof ModelListType];